% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treedater0.R
\name{goodnessOfFitPlot}
\alias{goodnessOfFitPlot}
\title{Produce a goodness of fit plot}
\usage{
goodnessOfFitPlot(td)
}
\arguments{
\item{td}{A treedater object generated by the \code{dater} function}
}
\description{
The sorted tail probabilties (p values) for each edge in the tree under the fitted model
}
