% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers0.R
\name{outlierTips}
\alias{outlierTips}
\title{Detect terminal lineages with unusually large evolutionary divergence under the fitted treedater model}
\usage{
outlierTips(td, alpha = 0.05)
}
\arguments{
\item{td}{A fitted treedater object}

\item{alpha}{The tail probability used for classifying lineages as outliers}
}
\value{
A data frame summarizing for each lineage the p values, adjusted p values ('q'), likelihood, rates, and branch lengths.
}
\description{
This is a convient wrapper of the *outlier.lineages*
}
\seealso{
dater
outlier.lineages
}
