% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{plot.bootTreedater}
\alias{plot.bootTreedater}
\title{Plots lineages through time and confidence intervals estimated by bootstrap.}
\usage{
\method{plot}{bootTreedater}(x, t0 = NA, res = 100, ggplot = FALSE, ...)
}
\arguments{
\item{x}{A bootTreedater object produced by *parboot* or *boot*}

\item{t0}{The lower bound of the time axis to show}

\item{res}{The number of time points on the time axis}

\item{ggplot}{If TRUE, will return a plot made with the ggplot2 package}

\item{...}{Additional arg's are passed to *ggplot* or *plot*}
}
\description{
Plots lineages through time and confidence intervals estimated by bootstrap.
}
