% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{relaxedClockTest}
\alias{relaxedClockTest}
\title{Use parametric bootstrap to test if relaxed clock offers improved fit to data.}
\usage{
relaxedClockTest(..., nreps = 100, overrideTempConstraint = T)
}
\arguments{
\item{...}{arguments passed to *dater*}

\item{nreps}{Integer number of simulations}

\item{overrideTempConstraint}{see *parboot*}
}
\value{
A list with elements:
\itemize{
\item strict_treedater: A dater object under a strict clock
\item relaxed_treedater: A dater object under a relaxed clock
\item clock: The favoured clock model 
\item parboot: Result of call to *parboot* using fitted treedater and forcing a relaxed clock
\item nullHypothesis_coef_of_variation_CI: The null hypothesis CV
}
}
\description{
This function simulates phylogenies with branch lengths in units
     of substitutions per site. Simulations are based on a fitted
     treedater object which provides parameters of the molecular clock
     model. The coefficient of variation of rates is estimated using a
     relaxed clock model applied to strict clock simulations. Estimates
     of the CV is then compared to the null distribution provided by
     simulations.
}
\details{
This function will print the optimal clock model
     and the distribution of the coefficient of variation statistic under the null hypothesis (strict
     clock). Parameters passed to this function should be the same as when calling *dater*.
}
\author{
Erik M Volz <erik.volz@gmail.com>
}
